pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tda-of-digital-images"
version = "0.1.0"
description = "Topological data analysis for digital images; persistent homology transformer experiments."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.0",
    "torchvision",
    "torchmetrics",
    "torcheval",
    "numpy",
    "scikit-learn",
    "pandas",
    "matplotlib",
    "seaborn",
    "tokenizers",
    "transformers",
    "datasets",
    "evaluate",
    "ripser",
    "persim",
    "gudhi",
    "tqdm",
    "mlflow",
    "pyyaml",
]

[project.optional-dependencies]
jupyter = ["jupyter"]
mof = ["coremof-tools", "pymatgen", "pyzeo"]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.build.targets.sdist]
exclude = ["src/test_datasets.ipynb"]
